Recommend usage of `require: rubocop-rails-accessibility`; rename RuboCop department to "RailsAccessibility"
