+++ guides/image-has-alt.md
@@ -1,4 +1,4 @@
-# RubocopRailsAccessibility/ImageHasAlt
+# RailsAccessibility/ImageHasAlt
 
 ## Rule Details
