--- lib/rubocop/cop/rubocop_rails_accessibility_cops.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require "pathname"
-require "yaml"
-
-require_relative "inject"
-require_relative "version"
-require_relative "rubocop-rails-accessibility/image_has_alt"
-require_relative "rubocop-rails-accessibility/no_positive_tabindex"
-require_relative "rubocop-rails-accessibility/no_redundant_image_alt"
