--- lib/rubocop/cop/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module RubocopRailsAccessibility
-  VERSION = "0.1.2"
-end
