+++ lib/rubocop/rails_accessibility/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RuboCop
+  module RailsAccessibility
+    VERSION = "0.1.2"
+  end
+end
